Cleanup K/s display to show only tenths, also make sure we use floating point math

+++ trunk/src/plugin.cpp
@@ -535,8 +535,8 @@ int32 nsPluginInstance::Write(NPStream * stream, int32 offset, int32 len, void *
 
                     // send_signal_with_double(this, item, "SetPercent", percent);
                     send_signal_with_double(this, item, "SetCachePercent", percent);
-                    rate = ((item->localsize - item->lastsize) / 1024) / difftime(time(NULL), lastupdate);
-                    text = g_strdup_printf(_("Cache fill: %2.2f%% (%2.2f K/s)"), percent * 100.0, rate);
+                    rate = (gdouble)((item->localsize - item->lastsize) / 1024.0) / (gdouble)difftime(time(NULL), lastupdate);
+                    text = g_strdup_printf(_("Cache fill: %2.2f%% (%0.1f K/s)"), percent * 100.0, rate);
                     send_signal_with_string(this, item, "SetProgressText", text);
                 }
                 time(&lastupdate);
